Implemented card animation, Created some Object hierarchy classes

+++ Blackjack/Source/Core/Private/Core/Utils.cpp
@@ -7,13 +7,13 @@ namespace Core
 	float Time::GetTime()
 	{
 		  Uint64 currentTime = SDL_GetPerformanceCounter();
-		  return static_cast<float>((currentTime - startTime) / SDL_GetPerformanceFrequency());
+		  return static_cast<float>((float)(currentTime - startTime) / (float)SDL_GetPerformanceFrequency());
 	}
 
 	float Time::GetTimeMillis()
 	{
 		Uint64 currentTime = SDL_GetPerformanceCounter();
-		return static_cast<float>((currentTime - startTime) * 1000 / SDL_GetPerformanceFrequency());
+		return static_cast<float>((float)(currentTime - startTime) * 1000 / (float)SDL_GetPerformanceFrequency());
 	}
 }
